Move library code from iree/tools/ to compiler/.
