+++ compiler/src/iree/compiler/API/Compiler.cpp
@@ -12,11 +12,11 @@
 #include "iree/compiler/InputConversion/MHLO/Passes.h"
 #include "iree/compiler/InputConversion/TOSA/Passes.h"
 #include "iree/compiler/Pipelines/Pipelines.h"
+#include "iree/compiler/Tools/init_dialects.h"
+#include "iree/compiler/Tools/init_llvmir_translations.h"
+#include "iree/compiler/Tools/init_passes.h"
+#include "iree/compiler/Tools/init_targets.h"
 #include "iree/compiler/Utils/OptionUtils.h"
-#include "iree/tools/init_dialects.h"
-#include "iree/tools/init_llvmir_translations.h"
-#include "iree/tools/init_passes.h"
-#include "iree/tools/init_targets.h"
 #include "mlir/CAPI/IR.h"
 #include "mlir/CAPI/Pass.h"
 #include "mlir/CAPI/Support.h"
