+++ compiler/src/iree/compiler/Tools/init_targets.cc
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/init_targets.h"
+#include "iree/compiler/Tools/init_targets.h"
 
 #include <functional>
