+++ compiler/src/iree/compiler/Tools/iree_compile_lib.cc
@@ -12,13 +12,13 @@
 #include "iree/compiler/ConstEval/Passes.h"
 #include "iree/compiler/Dialect/VM/Target/init_targets.h"
 #include "iree/compiler/Pipelines/Pipelines.h"
+#include "iree/compiler/Tools/init_dialects.h"
+#include "iree/compiler/Tools/init_llvmir_translations.h"
+#include "iree/compiler/Tools/init_passes.h"
+#include "iree/compiler/Tools/init_targets.h"
+#include "iree/compiler/Tools/iree_translate_lib.h"
 #include "iree/compiler/Utils/PassUtils.h"
 #include "iree/compiler/Utils/TracingUtils.h"
-#include "iree/tools/init_dialects.h"
-#include "iree/tools/init_llvmir_translations.h"
-#include "iree/tools/init_passes.h"
-#include "iree/tools/init_targets.h"
-#include "iree/tools/iree_translate_lib.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/InitLLVM.h"
