+++ iree/tools/iree-compile-main.cc
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/iree_translate_lib.h"
+#include "iree/compiler/Tools/iree_translate_lib.h"
 
 int main(int argc, char **argv) {
   return mlir::iree_compiler::runIreecMain(argc, argv);
